Refactor server draft files
