+++ server/draftTool/phaseHandlers/banPhase1Handler.ts
@@ -1,22 +1,24 @@
-import { HandlerVarsProps } from "./draftState";
-import { updateClientState } from "./clientDraftState";
+import { HandlerVarsProps } from "../states/draftState";
+import { updateClientState } from "../states/clientDraftState";
 
-export const banPhase2Handler = async ({
+export const banPhase1Handler = async ({
   io,
   lobbyCode,
   state,
   emitter,
 }: HandlerVarsProps): Promise<boolean> => {
-  if (state.activePhase !== "banPhase2") {
+  if (state.activePhase !== "banPhase1") {
     return false;
   }
 
   return new Promise((resolve, reject) => {
-    const bansPhase2 = [
-      state.redUser,
+    const bansPhase1 = [
       state.blueUser,
       state.redUser,
       state.blueUser,
+      state.redUser,
+      state.blueUser,
+      state.redUser,
     ];
 
     const startBanPhase = async () => {
@@ -25,10 +27,10 @@ export const banPhase2Handler = async ({
 
       for (
         state.banIndex;
-        state.banIndex < bansPhase2.length;
+        state.banIndex < bansPhase1.length;
         state.banIndex++
       ) {
-        const currentSide = bansPhase2[state.banIndex];
+        const currentSide = bansPhase1[state.banIndex];
         state.currentTurn = currentSide;
         try {
           // Display Current Turn in Client
@@ -60,7 +62,6 @@ export const banPhase2Handler = async ({
           timer--;
           state.timer = timer;
           io.to(lobbyCode).emit("timer", timer);
-
           if (timer <= 0) {
             clearInterval(interval);
             if (currentSide === state.blueUser) {
@@ -104,7 +105,7 @@ export const banPhase2Handler = async ({
               state.bansArray.push(state.bluePick);
               io.to(lobbyCode).emit("setBan", updateClientState(lobbyCode));
               state.bluePick = null;
-              state.currentBlueBan++
+              state.currentBlueBan++;
               // Shut of listener incase it still is attached
               emitter.off("bluePick", banListener);
               emitter.off("redPick", banListener);
@@ -117,7 +118,7 @@ export const banPhase2Handler = async ({
               state.bansArray.push(state.redPick);
               io.to(lobbyCode).emit("setBan", updateClientState(lobbyCode));
               state.redPick = null;
-              state.currentRedBan++
+              state.currentRedBan++;
               // Shut of listener incase it still is attached
               emitter.off("bluePick", banListener);
               emitter.off("redPick", banListener);
